started work on a docker driver
with the docker driver you can build your elixir application on your local docker,
it will spin up a container and keep it alive during the process. It will mount the
local directory as a volume but then clone the latest commit into the build_path to
ensure no unsaved changes are included.

this required some changes in the driver behaviour to manage how the code is sent and
how scripts are executed.

there is still some work ahead, I think it should be possible to have custom driver
configuration (optionally) and the docker contains dont exit at all right now.