added new StopRelease step before Unarchive for deployments
I noticed with the most regular and simple mix releases for my test project that it sometimes
did not properly restart the application. By default mix release will generate a cookie file
which ends up in `releases/COOKIE` - however it is different for every new release unless you
configure it yourself.

By stopping it before unarchiving the new version you ensure stop uses the old COOKIE value
to communicate with the running service.

Also added a test to the bash script to ensure it is actually stopped/started, so you know
when something goes wrong.