preparing support for different storage options
in the current version it will always store the release tarball on your local machine,
with the upcoming changes it should be possible to use alternative storage options. For
example AWS S3.

This commit adds some early support for this, defaulting to local storage to keep behaviour
similar for existing users.