added ability to run pre_deploy and post_deploy scripts
This introduces new functionality to Mate to run certain commands on your application
either pre or post deploy. This could be used for database migrations for example.

I am not yet sure this is the cleanest way to implement this, so I might revise it later
but I am going to play with it in a personal project for a while to see what works and
doesnt